Add additional communication protocols for the equivalence experiments.
These are 7 communication protocols that all behave as buffers:
ABP, ABP (BW), CABP and Par all behave as a one-place buffer,
Onebit behaves as a two-place buffer,
SWP with window size n behaves as a buffer with capacity 2*n.

It is expected that for strong-bisimulation the PBESs are easy (the equivalence is trivially violated),
whereas the other equivalences are expected to hold.